requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-phone-field==1.8.1
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
Levenshtein==0.23.0
packaging==23.2
phonenumbers==8.13.26
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
